requirements.txt
adbutils==0.6.1
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
Deprecated==1.2.6
deprecation==2.0.7
humanize==0.5.1
idna==2.8
imageio==2.5.0
kiwisolver==1.1.0
logzero==1.5.0
lxml==4.4.1
matplotlib==3.1.1
networkx==2.3
numpy==1.17.2
opencv-python==3.4.1.15
packaging==19.2
Pillow==6.1.0
progress==1.5
py==1.8.0
pyHook==1.5.1
pynput==1.4.4
pyparsing==2.4.2
python-dateutil==2.8.0
PyUserInput==0.1
PyWavelets==1.0.3
pywin32==225
requests==2.22.0
retry==0.9.2
scikit-image==0.15.0
scipy==1.3.1
six==1.12.0
uiautomator2==1.3.6
urllib3==1.25.6
whichcraft==0.6.1
wincertstore==0.2
wrapt==1.11.1
